Task11_update1

+++ restaurants/views.py
@@ -59,6 +59,8 @@ def restaurant_detail(request, restaurant_id):
     return render(request, 'detail.html', context)
 
 def restaurant_create(request):
+    if request.user.is_anonymous:
+        return redirect('signin')
     form = RestaurantForm()
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES)
@@ -75,6 +77,8 @@ def restaurant_create(request):
 def item_create(request, restaurant_id):
     form = ItemForm()
     restaurant = Restaurant.objects.get(id=restaurant_id)
+    if restaurant.owner!=request.user and not request.user.is_staff:
+        return redirect('noaccess')
     if request.method == "POST":
         form = ItemForm(request.POST)
         if form.is_valid():
@@ -90,6 +94,8 @@ def item_create(request, restaurant_id):
 
 def restaurant_update(request, restaurant_id):
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
+    if restaurant_obj.owner!=request.user and not request.user.is_staff:
+        return redirect('noaccess')
     form = RestaurantForm(instance=restaurant_obj)
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES, instance=restaurant_obj)
@@ -103,6 +109,14 @@ def restaurant_update(request, restaurant_id):
     return render(request, 'update.html', context)
 
 def restaurant_delete(request, restaurant_id):
+    if not request.user.is_staff:
+        return redirect('noaccess')
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
     restaurant_obj.delete()
     return redirect('restaurant-list')
+
+def noaccess(request):
+    context={
+        "msg":"you do not have access"
+    }
+    return render(request,'noaccess.html',context)
